+++ src/main/java/de/retest/recheck/review/ignore/AttributeFilter.java
@@ -1,11 +1,19 @@
 package de.retest.recheck.review.ignore;
 
+import java.util.Arrays;
 import java.util.Optional;
 import java.util.regex.MatchResult;
 import java.util.regex.Pattern;
 
+import org.apache.commons.lang3.tuple.Pair;
+
+import de.retest.recheck.ignore.AllMatchFilter;
+import de.retest.recheck.ignore.AllMatchFilter.AllMatchFilterLoader;
 import de.retest.recheck.ignore.Filter;
 import de.retest.recheck.review.ignore.AttributeRegexFilter.AttributeRegexFilterLoader;
+import de.retest.recheck.review.ignore.PixelDiffFilter.PixelDiffFilterLoader;
+import de.retest.recheck.review.ignore.io.InheritanceLoader;
+import de.retest.recheck.review.ignore.io.Loader;
 import de.retest.recheck.review.ignore.io.RegexLoader;
 import de.retest.recheck.ui.descriptors.Element;
 import lombok.extern.slf4j.Slf4j;
@@ -34,10 +42,13 @@ public class AttributeFilter implements Filter {
 	}
 
 	@Slf4j
-	public static class AttributeFilterLoader extends RegexLoader<AttributeFilter> {
+	public static class AttributeFilterLoader extends RegexLoader<Filter> {
 
-		static final String KEY = "attribute=";
+		public static final Loader<Filter> chainableFilter = new InheritanceLoader<>( Arrays.asList( //
+				Pair.of( PixelDiffFilter.class, new PixelDiffFilterLoader() ) //
+		) );
 
+		private static final String KEY = "attribute=";
 		private static final String FORMAT = KEY + "%s";
 		private static final Pattern REGEX = Pattern.compile( KEY + "(.+)" );
 
@@ -53,15 +64,26 @@ public class AttributeFilter implements Filter {
 		}
 
 		@Override
-		protected Optional<AttributeFilter> load( final MatchResult regex ) {
-			final String attribute = regex.group( 1 );
+		public String save( final Filter ignore ) {
+			if ( ignore instanceof AllMatchFilter ) {
+				return new AllMatchFilterLoader().save( (AllMatchFilter) ignore );
+			}
+			return super.save( ignore );
+		}
+
+		@Override
+		protected Optional<Filter> load( final MatchResult regex ) {
+			return ChainableFilterLoaderUtil.load( regex, this::loadSimpleAttributeFilter );
+		}
+
+		private AttributeFilter loadSimpleAttributeFilter( final String attribute ) {
 			if ( attribute.contains( POSSIBLE_REGEX ) ) {
 				final String actualLine = KEY + attribute;
 				final String suggestedLine = AttributeRegexFilterLoader.KEY + attribute;
 				log.warn( "'{}' contains '{}'. For regular expressions, please use '{}'.", actualLine, POSSIBLE_REGEX,
 						suggestedLine );
 			}
-			return Optional.of( new AttributeFilter( attribute.trim() ) );
+			return new AttributeFilter( attribute.trim() );
 		}
 	}